requirements.txt
art==6.4
color-it==2.1.3
pillow==11.0.0
sty==1.0.6
